Initial support for adding & removing maps
BlueMap's default maps folder becomes the map-templates folder, instead, and a new empty maps folder gets created.
Users can click on the "New map" button to create a new map config from one of those templates.

There is also a delete button which closes the config file in the editor if it's open, and then deletes the file the next frame.

The config tree UI is updated through a file watcher.